路径创建与权限校验对外服务接口实现 与提供feign调用的sdk， permission-sdk pom.xml增加 common模块

+++ sh-coupon-service/coupon-permission-service/src/main/java/com/sh/coupon/dao/UserRoleMappingRepository.java
@@ -13,5 +13,5 @@ public interface UserRoleMappingRepository extends JpaRepository<UserRoleMapping
      * @param userId
      * @return
      */
-    UserRoleMapping findByUserId(Integer userId);
+    UserRoleMapping findByUserId(Long userId);
 }
